backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg[binary]==3.2.3
python-dotenv==1.0.0
requests==2.31.0
